Natural_Language_Official/Natural_Language_Django/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
Pillow==10.4.0
numpy==2.0.2
matplotlib==3.9.2
scipy==1.13.1
sounddevice==0.5.1